requirements.txt
streamlit
transformers
torch
sentencepiece
langdetect
pandas
google-genai
